MPS-26410 language extensions can't be used in constrains and typesystem aspect: To address smooth transition, do not turn GP for aspects models when they use languages we have no idea about, as these may not get well along with MPS defined GP.

+++ languages/languageDesign/structure/source_gen/jetbrains/mps/lang/structure/migration/UseDevkit.java
@@ -8,8 +8,11 @@ import org.jetbrains.mps.openapi.module.SModule;
 import org.jetbrains.mps.openapi.model.SModel;
 import jetbrains.mps.lang.smodel.generator.smodelAdapter.SModuleOperations;
 import jetbrains.mps.smodel.SModelInternal;
-import jetbrains.mps.smodel.adapter.structure.MetaAdapterFactory;
+import org.jetbrains.mps.openapi.module.SModuleReference;
 import org.jetbrains.mps.openapi.persistence.PersistenceFacade;
+import java.util.function.Predicate;
+import org.jetbrains.mps.openapi.language.SLanguage;
+import jetbrains.mps.smodel.adapter.structure.MetaAdapterFactory;
 import jetbrains.mps.lang.migration.runtime.base.MigrationScriptReference;
 
 public class UseDevkit extends MigrationScriptBase {
@@ -30,12 +33,25 @@ public class UseDevkit extends MigrationScriptBase {
       return;
     }
     SModelInternal mi = (SModelInternal) structureAspectModel;
+    final SModuleReference structAspectDevkit = PersistenceFacade.getInstance().createModuleReference("78434eb8-b0e5-444b-850d-e7c4ad2da9ab(jetbrains.mps.devkit.aspect.structure)");
+    if (mi.importedDevkits().contains(structAspectDevkit)) {
+      return;
+    }
+    if (mi.importedLanguageIds().stream().anyMatch(new Predicate<SLanguage>() {
+      public boolean test(SLanguage l) {
+        return !(l.getQualifiedName().startsWith("jetbrains.mps."));
+      }
+    })) {
+      // Transition code, in case aspect uses custom extensions, do not turn GP on for it yet. 
+      return;
+    }
+
     mi.deleteLanguageId(MetaAdapterFactory.getLanguage(0xc72da2b97cce4447L, 0x8389f407dc1158b7L, "jetbrains.mps.lang.structure"));
     mi.deleteLanguageId(MetaAdapterFactory.getLanguage(0x982eb8df2c964bd7L, 0x996311712ea622e5L, "jetbrains.mps.lang.resources"));
     // sometimes there are devkits either too broad for the task (language design) or unrelated (general purpose is about BL), drop them 
     mi.deleteDevKit(PersistenceFacade.getInstance().createModuleReference("fbc25dd2-5da4-483a-8b19-70928e1b62d7(jetbrains.mps.devkit.general-purpose)"));
     mi.deleteDevKit(PersistenceFacade.getInstance().createModuleReference("2677cb18-f558-4e33-bc38-a5139cee06dc(jetbrains.mps.devkit.language-design)"));
-    mi.addDevKit(PersistenceFacade.getInstance().createModuleReference("78434eb8-b0e5-444b-850d-e7c4ad2da9ab(jetbrains.mps.devkit.aspect.structure)"));
+    mi.addDevKit(structAspectDevkit);
   }
   public MigrationScriptReference getDescriptor() {
     return new MigrationScriptReference(MetaAdapterFactory.getLanguage(0xc72da2b97cce4447L, 0x8389f407dc1158b7L, "jetbrains.mps.lang.structure"), 3);
